+++ src/main/java/kitchenpos/dao/JdbcTemplateMenuDao.java
@@ -1,6 +1,6 @@
 package kitchenpos.dao;
 
-import kitchenpos.domain.Menu;
+import kitchenpos.menu.domain.Menu;
 import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
 import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
@@ -70,10 +70,10 @@ public class JdbcTemplateMenuDao implements MenuDao {
 
     private Menu toEntity(final ResultSet resultSet) throws SQLException {
         final Menu entity = new Menu();
-        entity.setId(resultSet.getLong("id"));
-        entity.setName(resultSet.getString("name"));
-        entity.setPrice(resultSet.getBigDecimal("price"));
-        entity.setMenuGroupId(resultSet.getLong("menu_group_id"));
+//        entity.setId(resultSet.getLong("id"));
+//        entity.setName(resultSet.getString("name"));
+//        entity.setPrice(resultSet.getBigDecimal("price"));
+//        entity.setMenuGroupId(resultSet.getLong("menu_group_id"));
         return entity;
     }
 }
