+++ src/main/java/kitchenpos/dao/JdbcTemplateMenuGroupDao.java
@@ -1,6 +1,6 @@
 package kitchenpos.dao;
 
-import kitchenpos.domain.MenuGroup;
+import kitchenpos.menu.domain.MenuGroup;
 import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
 import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
@@ -70,8 +70,8 @@ public class JdbcTemplateMenuGroupDao implements MenuGroupDao {
 
     private MenuGroup toEntity(final ResultSet resultSet) throws SQLException {
         final MenuGroup entity = new MenuGroup();
-        entity.setId(resultSet.getLong("id"));
-        entity.setName(resultSet.getString("name"));
+//        entity.setId(resultSet.getLong("id"));
+//        entity.setName(resultSet.getString("name"));
         return entity;
     }
 }
