+++ src/main/java/kitchenpos/dao/JdbcTemplateOrderDao.java
@@ -97,8 +97,8 @@ public class JdbcTemplateOrderDao implements OrderDao {
     private Order toEntity(final ResultSet resultSet) throws SQLException {
         final Order entity = new Order();
         entity.setId(resultSet.getLong(KEY_COLUMN_NAME));
-        entity.setOrderTableId(resultSet.getLong("order_table_id"));
-        entity.setOrderStatus(resultSet.getString("order_status"));
+//        entity.setOrderTableId(resultSet.getLong("order_table_id"));
+//        entity.setOrderStatus(resultSet.getString("order_status"));
         entity.setOrderedTime(resultSet.getObject("ordered_time", LocalDateTime.class));
         return entity;
     }
