+++ src/main/java/kitchenpos/dao/JdbcTemplateProductDao.java
@@ -1,6 +1,6 @@
 package kitchenpos.dao;
 
-import kitchenpos.domain.Product;
+import kitchenpos.product.domain.Product;
 import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
 import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
@@ -64,7 +64,7 @@ public class JdbcTemplateProductDao implements ProductDao {
         final Product entity = new Product();
         entity.setId(resultSet.getLong(KEY_COLUMN_NAME));
         entity.setName(resultSet.getString("name"));
-        entity.setPrice(resultSet.getBigDecimal("price"));
+//        entity.setPrice(resultSet.getBigDecimal("price"));
         return entity;
     }
 }
